requirements.txt
streamlit==1.32.2
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
azure-cognitiveservices-speech==1.37.0
pydub==0.25.1
loguru==0.7.2
tenacity==8.2.3